+++ MapboxSearch/autofill/src/main/java/com/mapbox/search/autofill/AddressAutofillResult.kt
@@ -1,6 +1,7 @@
 package com.mapbox.search.autofill
 
 import android.os.Parcelable
+import com.mapbox.geojson.Point
 import kotlinx.parcelize.Parcelize
 
 /**
@@ -25,6 +26,11 @@ public class AddressAutofillResult internal constructor(
     public val suggestion: AddressAutofillSuggestion,
 
     /**
+     * Place geographic point.
+     */
+    public val coordinate: Point,
+
+    /**
      * Detailed address components like street, house number, etc.
      */
     public val address: AddressComponents,
@@ -42,6 +48,7 @@ public class AddressAutofillResult internal constructor(
         if (id != other.id) return false
         if (mapboxId != other.mapboxId) return false
         if (suggestion != other.suggestion) return false
+        if (coordinate != other.coordinate) return false
         if (address != other.address) return false
 
         return true
@@ -54,6 +61,7 @@ public class AddressAutofillResult internal constructor(
         var result = id.hashCode()
         result = 31 * result + mapboxId.hashCode()
         result = 31 * result + suggestion.hashCode()
+        result = 31 * result + coordinate.hashCode()
         result = 31 * result + address.hashCode()
         return result
     }
@@ -62,6 +70,6 @@ public class AddressAutofillResult internal constructor(
      * @suppress
      */
     override fun toString(): String {
-        return "AddressAutofillResult(id=$id, mapboxId=$mapboxId, suggestion=$suggestion, address=$address)"
+        return "AddressAutofillResult(id=$id, mapboxId=$mapboxId, suggestion=$suggestion, coordinate=$coordinate, address=$address)"
     }
 }