+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## 1+
 
 - [CORE] Added a new field called `mapboxId` to `BaseSearchResult`, `SearchResult`, `CategoryResult` and `OfflineSearchResult`
+- [Address Autofill, Place Autocomplete] Search is a two-steps action now, i.e. it returns `Suggestion`s (without the geo coordinate) at the first step and `Result` after suggestion selection. The `coordinate` field is no longer available in `Suggestion`.
+- [Address Autofill] `suggestions()` is renamed to `reverseGeocoding()`
 
 ## 1.0.0-rc.8